--- src/test/java/io/github/eocqrs/eokson/SmartJsonBoolLeafTest.java
@@ -1,215 +1,0 @@
-/*
- *  Copyright (c) 2023 Aliaksei Bialiauski, EO-CQRS
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NON-INFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-package io.github.eocqrs.eokson;
-
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.node.ObjectNode;
-import io.github.eocqrs.eokson.Json;
-import io.github.eocqrs.eokson.SmartJson;
-import org.junit.jupiter.api.Test;
-
-final class SmartJsonBoolLeafTest {
-
-    private static final ObjectMapper MAPPER = new ObjectMapper();
-
-    // Immediate, normal
-
-    @Test
-    void findsOptLeafAsBool() {
-        assertTrue(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .put("field2", true)
-                )
-            ).optLeafAsBool("field2").get()
-        );
-    }
-
-    @Test
-    void findsLeafAsBool() {
-        assertTrue(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .put("field2", true)
-                )
-            ).leafAsBool("field2")
-        );
-    }
-
-    @Test
-    void returnsEmptyOptionalForNonexistentBooleanLeaf() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                )
-            ).optLeafAsBool("nonexistent").isPresent()
-        );
-    }
-
-    @Test
-    void throwsForNonexistentBooleanLeaf() {
-        assertTrue(
-            assertThrows(
-                IllegalArgumentException.class,
-                () -> new SmartJson(
-                    new Json.Of(
-                        MAPPER.createObjectNode()
-                    )
-                ).leafAsBool("nonexistent")
-            ).getMessage().contains("No such field")
-        );
-    }
-
-    // Immediate, autoconversion
-
-    @Test
-    void returnsFalseIfOptLeafIsNotBool() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("stringField", "value1")
-                        .put("boolField", true)
-                )
-            ).optLeafAsBool("stringField").get()
-        );
-    }
-
-    @Test
-    void returnsFalseIfLeafIsNotBool() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("stringField", "value1")
-                        .put("boolField", true)
-                )
-            ).leafAsBool("stringField")
-        );
-    }
-
-    // Path, normal
-
-    @Test
-    void findsOptBoolLeafInPath() {
-        assertTrue(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .<ObjectNode>set(
-                            "field2",
-                            MAPPER.createObjectNode()
-                                .put("innerField", true)
-                        )
-                )
-            ).optLeafAsBool("/field2/innerField").get()
-        );
-    }
-
-    @Test
-    void findsLeafBoolInPath() {
-        assertTrue(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .<ObjectNode>set(
-                            "field2",
-                            MAPPER.createObjectNode()
-                                .put("innerField", true)
-                        )
-                )
-            ).leafAsBool("/field2/innerField")
-        );
-    }
-
-    @Test
-    void returnsEmptyOptionalForNonexistentLeafBoolInPath() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                )
-            ).optLeafAsBool("/nonexistent/path").isPresent()
-        );
-    }
-
-    @Test
-    void throwsForNonexistentLeafBoolInPath() {
-        assertTrue(
-            assertThrows(
-                IllegalArgumentException.class,
-                () -> new SmartJson(
-                    new Json.Of(
-                        MAPPER.createObjectNode()
-                    )
-                ).leafAsBool("/nonexistent/path")
-            ).getMessage().contains("No such field")
-        );
-    }
-
-    // Path, autoconversion
-
-    @Test
-    void returnsFalseIfOptLeafInPathIsNotBool() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .<ObjectNode>set(
-                            "field2",
-                            MAPPER.createObjectNode()
-                                .put("innerField", "string")
-                        )
-                )
-            ).optLeafAsBool("/field2/innerField").get()
-        );
-    }
-
-    @Test
-    void returnsFalseIfLeafInPathIsNotBool() {
-        assertFalse(
-            new SmartJson(
-                new Json.Of(
-                    MAPPER.createObjectNode()
-                        .put("field1", "value1")
-                        .<ObjectNode>set(
-                            "field2",
-                            MAPPER.createObjectNode()
-                                .put("innerField", 17)
-                        )
-                )
-            ).leafAsBool("/field2/innerField")
-        );
-    }
-}
